fix: Improve test infrastructure and database maintenance service
- Add explicit save operation to DatabaseMaintenanceService after maintenance operations
- Enhance DTXAPIClientTests with proper UserDefaults synchronization and verification
- Refactor MetronomeBasicTests to use async/await patterns for reliable timing tests
- Split large TestInfrastructure.swift file to satisfy SwiftLint length limits:
  - Move TestHelpers and AsyncTestingUtilities to TestHelpers.swift
  - Move TestModelFactory to separate TestModelFactory.swift
  - Keep core infrastructure in TestInfrastructure.swift
- Improve TestAssertions deletion checks with proper error handling for SwiftData edge cases
- Fix test flakiness by using proper async waiting patterns instead of synchronous assertions
